Fix ADXL362 driver bugs found while processing PR #164
- Clean up ~40 untracked, uncommitted stray files (duplicate AD7705
  content and empty Cargo.toml/main.cpp stubs) left sitting in the
  working tree under go/examples/*/accelerometer/adxl362/complete/ and
  equivalent cpp/rust paths — these were never part of the PR's git
  history but were confusing go vet/cargo build with bogus targets
- Go: fix a uint16/uint8 type-mismatch comparison in ReadFifo(); fix
  two Go string literals using invalid leading-`+` line continuation
  (concatenation must break after the operator, not before it) in the
  Linux and TinyGo demo examples; discard the unused `revid` return
  value in the hardware test instead of binding it
- Rust: fix two example Cargo.toml files both named "adxl362_minimal"
  (copy-paste — should be adxl362_complete/adxl362_demo); fix the same
  invalid string continuation in the Linux demo (merged into single
  literals, since Rust doesn't concatenate adjacent string literals);
  fix a "literal out of range for u8" bug in read_8bit() using an
  unneeded, buggy manual sign-extension when `v as i8` already does
  the correct two's-complement reinterpretation cast; alias the
  ADXL362 FIFO_STREAM re-export in accelerometer/mod.rs (identical
  name, different value, to ADXL345's own FIFO_STREAM) to
  ADXL362_FIFO_STREAM, matching the project's established `as` import-
  alias convention for cross-chip name collisions in the same category
- Kotlin: add missing `@JvmStatic` on two protected non-const companion
  functions (signExtend12, sleepMs) that Adxl362Full calls — the same
  class of bug fixed in the AD7705/AD7706 PRs; fix a Float/Double
  argument-type mismatch in read8bit()
- Groovy: fix "possible loss of precision from long to int" from
  Math.round(double) under @CompileStatic by adding the (int) cast
  used everywhere else in this codebase for the same pattern
- Regenerate cpp/README.md and the two Node.js package READMEs
- sigrok/adxl362/pd.py: fix a real decoding bug found while adding the
  missing conformance checker (see below) — the decoder's read-path
  state machine unconditionally captured dummy MOSI bytes instead of
  the real MISO data for every register read (STATUS, FILTER_CTL,
  POWER_CTL, etc. were all decoding garbage), because it set
  state='GET_DATA_WRITE' after the register-pointer byte regardless of
  whether the command was a read or a write. Also: FIFO reads emitted
  their annotation immediately on seeing the 0x0D command byte, before
  any FIFO data had actually been captured, so they always produced
  zero entries. Both fixed and verified against synthetic decode()
  event sequences (register read incl. multi-byte burst, FIFO read,
  register write, SOFT_RESET) since no real capture hardware is
  available here. Added the XDATA_L..ZDATA_H/TEMP_L-H burst-read
  reconstruction the module's own docstring already claimed to support
  but never implemented (only a `pass`-stub existed), and the
  data_read_start/data_read_done annotation pair the spec's own Sigrok
  Decoder section names for the data-ready-clear-latency check
- Add the two files the spec's Implementation Checklist and the
  AD7705/AD7706 precedent both call for but this PR omitted:
  conformance/accelerometer/adxl362_conformance.py (CHECKS
  intentionally empty/deferred, same reasoning as AD7705/AD7706 —
  interface parity only) and specs/accelerometer/adxl362_timing.conf

+++ go/periph/chips/accelerometer/adxl362.go
@@ -562,7 +562,7 @@ func (f *ADXL362Full) ReadFifo() ([]FifoEntry, error) {
 		axis := (raw16 >> 14) & 0x03
 		raw12 := signExtend12(raw16 & 0x0FFF)
 		var val float32
-		if axis == ADXL362AxisTemp {
+		if uint8(axis) == ADXL362AxisTemp {
 			val = 25.0 + (float32(raw12)-350)*0.065
 		} else {
 			val = float32(raw12) * sens
